+++ docs/source/conf.py
@@ -168,7 +168,8 @@ texinfo_documents = [
 ]
 
 # Example configuration for intersphinx: refer to the Python standard library.
-intersphinx_mapping = {'https://docs.python.org/3/': None}
+intersphinx_mapping = {'https://docs.python.org/3/': None,
+                       'http://wrapt.readthedocs.io/en/latest/': None}
 
 # -- Options for EPub output -------------------------------------------